Update analysis scripts to work with PISA
