+++ analysis/download_comparisons.py
@@ -7,6 +7,7 @@ import tempfile
 import time
 
 from kaggle.api.kaggle_api_extended import KaggleApi
+from utils import kernels_output
 
 ERROR_OOM = 'Your notebook tried to allocate more memory than is available.'
 ERROR_TIMEOUT = 'Your notebook was stopped because it exceeded the max allowed execution duration.'
@@ -16,53 +17,63 @@ output_base_dir = Path("./comparison_results")
 
 # use temp folder to store results
 methods_to_nb_name = {
-    'bm25-pt': [
-        'benchmark-bm25-pt-sub-1m',
-        'benchmark-bm25-pt-cqadupstack',
-        'benchmark-bm25-pt-webis-touche2020',
-        'benchmark-bm25-pt-nq',
-        'benchmark-bm25-pt-msmarco',
-        'benchmark-bm25-pt-hotpotqa',
-        'benchmark-bm25-pt-dbpedia-entity',
-        'benchmark-bm25-pt-fever', 
-        'benchmark-bm25-pt-climate-fever', 
-    ],
-    'rank-bm25': [
-        'benchmark-rank-bm25-sub-1m',
-        'benchmark-rank-bm25-cqadupstack',
-        'benchmark-rank-bm25-nq',
-        'benchmark-rank-bm25-msmarco',
-        'benchmark-rank-bm25-hotpotqa',
-        'benchmark-rank-bm25-dbpedia-entity',
-        'benchmark-rank-bm25-fever', 
-        'benchmark-rank-bm25-climate-fever', 
-    ],
-    'bm25s': [
-        "comparing-bm25s-lucene-k1-0-9-b-0-4",
-        "comparing-bm25s-lucene",
-        "run-bm25s-k1-1-5-b-0-75",
-        "comparing-bm25s-bm25l",
-        "comparing-bm25s-bm25plus",
-        "comparing-bm25s-no-stopwords-no-stemmer",
-        "comparing-bm25s-no-stopwords",
-        "comparing-bm25s-no-stemmer",
-        "comparing-bm25s-robertson",
-        "comparing-bm25s-atire",
-        "comparing-bm25s-lucene",
-    ],
-    'elasticsearch': [
-        "run-elasticsearch-k1-1-5-b-0-75"
-    ],
-    'pyserini': [
-        'benchmark-pyserini-sub-1m',
-        'benchmark-pyserini-cqadupstack',
-        'benchmark-pyserini-nq',
-        'benchmark-pyserini-msmarco',
-        'benchmark-pyserini-hotpotqa', 
-        'benchmark-pyserini-dbpedia-entity', 
-        'benchmark-pyserini-fever',  
-        'benchmark-pyserini-climate-fever',  
-    ],
+    # 'bm25-pt': [
+    #     'xhlulu/benchmark-bm25-pt-sub-1m',
+    #     'xhlulu/benchmark-bm25-pt-cqadupstack',
+    #     'xhlulu/benchmark-bm25-pt-webis-touche2020',
+    #     'xhlulu/benchmark-bm25-pt-nq',
+    #     'xhlulu/benchmark-bm25-pt-msmarco',
+    #     'xhlulu/benchmark-bm25-pt-hotpotqa',
+    #     'xhlulu/benchmark-bm25-pt-dbpedia-entity',
+    #     'xhlulu/benchmark-bm25-pt-fever', 
+    #     'xhlulu/benchmark-bm25-pt-climate-fever', 
+    # ],
+    # 'rank-bm25': [
+    #     'xhlulu/benchmark-rank-bm25-sub-1m',
+    #     'xhlulu/benchmark-rank-bm25-cqadupstack',
+    #     'xhlulu/benchmark-rank-bm25-nq',
+    #     'xhlulu/benchmark-rank-bm25-msmarco',
+    #     'xhlulu/benchmark-rank-bm25-hotpotqa',
+    #     'xhlulu/benchmark-rank-bm25-dbpedia-entity',
+    #     'xhlulu/benchmark-rank-bm25-fever', 
+    #     'xhlulu/benchmark-rank-bm25-climate-fever', 
+    # ],
+    # 'bm25s': [
+    #     "xhlulu/comparing-bm25s-lucene-k1-0-9-b-0-4",
+    #     "xhlulu/comparing-bm25s-lucene",
+    #     "xhlulu/run-bm25s-k1-1-5-b-0-75",
+    #     "xhlulu/comparing-bm25s-bm25l",
+    #     "xhlulu/comparing-bm25s-bm25plus",
+    #     "xhlulu/comparing-bm25s-no-stopwords-no-stemmer",
+    #     "xhlulu/comparing-bm25s-no-stopwords",
+    #     "xhlulu/comparing-bm25s-no-stemmer",
+    #     "xhlulu/comparing-bm25s-robertson",
+    #     "xhlulu/comparing-bm25s-atire",
+    #     "xhlulu/comparing-bm25s-lucene",
+    # ],
+    # 'elasticsearch': [
+    #     "xhlulu/run-elasticsearch-k1-1-5-b-0-75"
+    # ],
+    # 'pyserini': [
+    #     'xhlulu/benchmark-pyserini-sub-1m',
+    #     'xhlulu/benchmark-pyserini-cqadupstack',
+    #     'xhlulu/benchmark-pyserini-nq',
+    #     'xhlulu/benchmark-pyserini-msmarco',
+    #     'xhlulu/benchmark-pyserini-hotpotqa', 
+    #     'xhlulu/benchmark-pyserini-dbpedia-entity', 
+    #     'xhlulu/benchmark-pyserini-fever',  
+    #     'xhlulu/benchmark-pyserini-climate-fever',  
+    # ],
+    "pisa": [
+        "smac2048/pisa-nq",
+        "smac2048/pisa-rest",
+        "smac2048/pisa-dbpedia-entity",
+        "smac2048/pisa-climate-fever",
+        "smac2048/pisa-hotpotqa",
+        "smac2048/pisa-fever",
+        "smac2048/pisa-msmarco",
+        "smac2048/pisa-cqadupstack",
+    ]
 }
 
 
@@ -71,11 +82,9 @@ api = KaggleApi()
 api.authenticate()
 
 for method, notebook_names in methods_to_nb_name.items():
-    for name in notebook_names:   
+    for notebook_name in notebook_names:   
         output_dir = output_base_dir / method
         output_dir.mkdir(parents=True, exist_ok=True)
-
-        notebook_name = f'{username}/{name}'
 
         try:
             status = api.kernels_status(notebook_name)
@@ -97,14 +106,14 @@ for method, notebook_names in methods_to_nb_name.items():
 
             # save the error message to the same output dir as the results, but prefix with 'error'
             t = time.strftime('%Y%m%d-%H%M%S')
-            error_file = output_dir / f'error-{name}-{t}.txt'
+            error_file = output_dir / f'error-{notebook_name.replace("/", "_")}-{t}.txt'
             with open(error_file, 'w') as f:
                 f.write(failure_message)
             continue
             
         temp_dir = tempfile.mkdtemp()
 
-        out_files = api.kernels_output(notebook_name, temp_dir)
+        out_files = kernels_output(api, notebook_name, temp_dir, allowed_patterns=['*.json'])
     
         print(f"Downloaded results for '{notebook_name}'")    
         # only save out *.json files to the output folder
