+++ analysis/requirements.txt
@@ -1,2 +1,3 @@
 kaggle
 tabulate
+pandas
